implement basic functionality of to-do list
