--- .gitignore
@@ -1,2 +1,0 @@
-scheduler.db
-runTests.txt
