+++ README.md
@@ -1,4 +1,5 @@
-# Файлы для итогового задания
+# Реализован базовый функционал 
+Файлы для итогового задания
 
 В директории `tests` находятся тесты для проверки API, которое должно быть реализовано в веб-сервере.
